+++ routes.js
@@ -3,8 +3,7 @@ const router = express.Router();
 const details = require('./data/order');
 const Joi = require('joi');
 // const validate = require('./services/validate');
-const loadStores = require('./stores');
-const process = require('./services/process');
+const loadStores = require('./data/stores');
 
 /**
  * @swagger
@@ -92,10 +91,6 @@ router.get('/', async function (req, res) {
         console.log('password inside routes.js', (await dbConfig).password);
         console.log('database inside routes.js', (await dbConfig).database);
 
-        // Save store POS db details in ENV
-        process.saveStoreValuesInProcessEnv((await dbConfig).ipAddress, (await dbConfig).port, (await dbConfig).username,
-            (await dbConfig).password, (await dbConfig).database);
-
         // validate.input(ordNumber, storeNumber);
         // validate request body against schema
         // const { error, value } = schema.validate(req.query, options);
@@ -107,6 +102,7 @@ router.get('/', async function (req, res) {
         // on success invoke function to get Order details
         const results = await details.getOrderByOrdNumStoreID(ordNumber);
 
+        console.debug(results.details);
         if (results.details != 0)
             res.json({ 'status': 'READY' });
